Replace the intervening pads - final commit.
For reasons that I don't understand, the connection e.g. of the dsu_break
to the dsubre_pad caused an error late in the synthesis - I set it to '0'
to simulate a button that is never pressed, but that interracted badly
somehow with the 'inpad generic'; perhaps for the pads to work the signal
needs to actually come from a button (specified in a different way
(electrically) in the UCF? Don't know. But skipping the pad and connecting
dsubre directly to dsui.break worked just fine.

For clk and reset the errors were different - about multiple input buffers
detected in series. Clearly, the ZestSC1 already had buffers on these
signals, so I just skipped the pads again.

But alas, this will probably be my final commit - since I've now reached a problem
I can't solve:

    Logic Utilization:
      Number of Slice Flip Flops:         2,654 out of  15,360   17%
      Number of 4 input LUTs:             9,767 out of  15,360   63%
    Logic Distribution:
      Number of occupied Slices:          5,687 out of   7,680   74%
        Number of Slices containing only related logic:   5,687 out of   5,687 100%
        Number of Slices containing unrelated logic:          0 out of   5,687   0%
          *See NOTES below for an explanation of the effects of unrelated logic.
      Total Number of 4 input LUTs:       9,877 out of  15,360   64%
        Number used as logic:             9,716
        Number used as a route-thru:        110
        Number used for Dual Port RAMs:       2
          (Two LUTs used per Dual Port RAM)
        Number used as Shift registers:      49

      The Slice Logic Distribution report is not meaningful if the design is
      over-mapped for a non-slice resource or if Placement fails.

      Number of bonded IOBs:                192 out of     173  110% (OVERMAPPED)
        IOB Flip Flops:                      85
      Number of RAMB16s:                     17 out of      24   70%
      Number of MULT18X18s:                   1 out of      24    4%
      Number of BUFGMUXs:                     5 out of       8   62%
      Number of DCMs:                         4 out of       4  100%
      Number of BSCANs:                       1 out of       1  100%

The dreaded 'design too big to fit' :-( message.

My FPGA board is apparently not big enough for a Leon3...

Damn.